+++ src/constants.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Shared constants.
@@ -7,8 +7,8 @@ The use of this module should be limited to cases where the constant is not bett
 another module or to resolve circular imports.
 """
 DEFAULT_BRANCH = "main"
-DOCUMENTATION_TAG = "upload-charm-docs/base-content"
-DISCOURSE_AHEAD_TAG = "upload-charm-docs/discourse-ahead-ok"
+DOCUMENTATION_TAG = "discourse-gatekeeper/base-content"
+DISCOURSE_AHEAD_TAG = "discourse-gatekeeper/discourse-ahead-ok"
 
 DOCUMENTATION_FOLDER_NAME = "docs"
 DOC_FILE_EXTENSION = ".md"