+++ src/reconcile.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Module for calculating required changes based on docs directory and navigation table."""
@@ -678,8 +678,9 @@ def index_page(
         The action to take for the index page.
     """
     table_contents = "\n".join(table_row.to_markdown(discourse.host) for table_row in table_rows)
+    # Strip any whitespace around file contents to avoid buildup of whitespace
     local_content = (
-        f"{index_module.get_content_for_server(index.local)}{NAVIGATION_TABLE_START}\n"
+        f"{index_module.get_content_for_server(index.local).strip()}\n{NAVIGATION_TABLE_START}\n"
         f"{table_contents}\n".strip()
     )