+++ tests/factories.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Factories for generating test data."""
@@ -337,7 +337,7 @@ class UserInputDiscourseFactory(
         abstract = False
 
     hostname = factory.Sequence(lambda n: f"discourse/{n}")
-    category_id = factory.Sequence(lambda n: f"{n}")
+    category_id = factory.Sequence(str)
     api_username = factory.Sequence(lambda n: f"discourse-test-user-{n}")
     api_key = factory.Sequence(lambda n: f"discourse-test-key-{n}")