+++ tests/unit/test_migration/__init__.py
@@ -1,4 +1,4 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-"""Unit tests for actions."""
+"""Unit tests for migartion module."""